Edit: Moved errors to infra layer
